+++ src/admin/components/elements/BlockEditor/menus/LinkMenu/index.tsx
@@ -1,9 +1,8 @@
-import { IconButton } from '@mui/material';
 import { Editor } from '@tiptap/react';
 import { Link } from 'lucide-react';
 import React, { useCallback, useState } from 'react';
-import { EditLinkPopover } from './EditLinkPopover.js';
 import { ToolbarButton } from '../../ToolbarButton/index.js';
+import { EditLinkPopover } from './EditLinkPopover.js';
 
 type Props = {
   editor: Editor;
